requirements.txt
python-telegram-bot[flask]
flask
gunicorn
requests
